debug v1.0.6 收藏名片

+++ app/Http/Controllers/Home/CardcaseController.php
@@ -90,12 +90,7 @@ class CardcaseController extends Controller
                 $err_code = 701; // 收藏失败
             }
         }
-//        dd(isset($err_code));
-        if (isset($err_code)) {
-            Config::set('global.ajax.err', $err_code);
-            Config::set('global.ajax.msg', config('global.msg.' . $err_code));
-            return Config::get('global.ajax');
-        }
+
 
         /* 查看数据库是否有数据 */
         $query = Cardcase::query();
@@ -106,18 +101,20 @@ class CardcaseController extends Controller
 
         /* ajax收藏 */
         if ($request->ajax()) {
-            if ($cardcase) { // 有，删除
-                if ($cardcase->delete()) {
-                    $err_code = 750; // 取消收藏成功
-                } else {
-                    $err_code = 751; // 取消收藏失败
-                }
-            } else { // 无，收藏
-                /* 添加名片到名片夹 */
-                if (Cardcase::create($data)) {
-                    $err_code = 700; // 收藏成功
-                } else {
-                    $err_code = 701; // 收藏失败
+            if (!isset($err_code)) {
+                if ($cardcase) { // 有，删除
+                    if ($cardcase->delete()) {
+                        $err_code = 750; // 取消收藏成功
+                    } else {
+                        $err_code = 751; // 取消收藏失败
+                    }
+                } else { // 无，收藏
+                    /* 添加名片到名片夹 */
+                    if (Cardcase::create($data)) {
+                        $err_code = 700; // 收藏成功
+                    } else {
+                        $err_code = 701; // 收藏失败
+                    }
                 }
             }
             Config::set('global.ajax.err', $err_code);
@@ -126,11 +123,13 @@ class CardcaseController extends Controller
         }
 
         /* url收藏 */
-        if ($request->isMethod('get')) {
-            if ($cardcase) { // 有，删除
-                $err_code = 702; // 收藏失败
-            } elseif (Cardcase::create($data)) {
-                $err_code = 700; // 收藏成功
+        if (!isset($err_code)) {
+            if ($request->isMethod('get')) {
+                if ($cardcase) { // 有，删除
+                    $err_code = 702; // 收藏失败
+                } elseif (Cardcase::create($data)) {
+                    $err_code = 700; // 收藏成功
+                }
             }
         }
         return redirect('cardcase')->with('info', config('global.msg.' . $err_code));
